+++ src/QuantitativeWorld.Text.Json/Angular/AngleBuilder.cs
@@ -4,8 +4,8 @@ namespace QuantitativeWorld.Text.Json.Angular
 {
     internal class AngleBuilder : ILinearQuantityBuilder<Angle, AngleUnit>
     {
-        private decimal? _turns;
-        private decimal? _value;
+        private double? _turns;
+        private double? _value;
         private AngleUnit? _unit;
 
         public AngleBuilder() { }
@@ -16,13 +16,13 @@ namespace QuantitativeWorld.Text.Json.Angular
             _unit = angle.Unit;
         }
 
-        public void SetBaseValue(decimal turns)
+        public void SetBaseValue(double turns)
         {
             _turns = turns;
             _value = null;
         }
 
-        public void SetValue(decimal value)
+        public void SetValue(double value)
         {
             _turns = null;
             _value = value;
@@ -35,8 +35,8 @@ namespace QuantitativeWorld.Text.Json.Angular
 
         public bool TryBuild(out Angle result, AngleUnit? defaultUnit = null)
         {
-            decimal? turns = _turns;
-            decimal? value = _value;
+            double? turns = _turns;
+            double? value = _value;
             AngleUnit? unit = _unit ?? defaultUnit;
 
             if (turns.HasValue)
