+++ src/QuantitativeWorld/WeightFactory.cs
@@ -4,7 +4,7 @@ namespace QuantitativeWorld
 {
     public class WeightFactory : ILinearQuantityFactory<Weight, WeightUnit>
     {
-        public Weight Create(decimal value, WeightUnit unit) =>
+        public Weight Create(double value, WeightUnit unit) =>
             new Weight(value, unit);
     }
 }
